third_party/WebKit/LayoutTests/fast/shapedetection/shapedetection-empty-input.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script>

// This test verifies FaceDetector.detect() with an empty HTMLImageElement.
async_test(function(t) {
  var image = new Image();
  var faceDetector = new FaceDetector();
  var tryFaceDetection = function() {
    faceDetector.detect(image)
        .then(faceDetectionResult => {
          assert_equals(faceDetectionResult.length, 0);
          t.done();
        })
        .catch(error => {
          assert_unreached("detect() rejected with error: " + error)
        });
  };
  image.onload = tryFaceDetection;
  image.onerror = tryFaceDetection;
  image.src = "";
}, "FaceDetector should resolve empty images with empty results.");

</script>
